+++ azure-communication-ui/azure-communication-ui/src/calling/java/com/azure/android/communication/ui/calling/di/DependencyInjectionContainer.kt
@@ -15,7 +15,7 @@ import com.azure.android.communication.ui.calling.presentation.manager.Lifecycle
 import com.azure.android.communication.ui.calling.presentation.manager.PermissionManager
 import com.azure.android.communication.ui.calling.presentation.navigation.NavigationRouter
 import com.azure.android.communication.ui.calling.redux.Store
-import com.azure.android.communication.ui.calling.redux.middleware.bluetooth.AudioSwitchingMiddleware
+import com.azure.android.communication.ui.calling.redux.middleware.audio.AudioSwitchingMiddleware
 import com.azure.android.communication.ui.calling.redux.middleware.handler.CallingMiddlewareActionHandler
 import com.azure.android.communication.ui.calling.redux.state.ReduxState
 import com.azure.android.communication.ui.calling.service.NotificationService
