+++ azure-communication-ui/azure-communication-ui/src/calling/java/com/azure/android/communication/ui/calling/presentation/manager/DeviceDetectionManager.kt
@@ -4,14 +4,9 @@
 package com.azure.android.communication.ui.calling.presentation.manager
 
 import android.app.Activity
-import android.content.Context
-import com.azure.android.communication.ui.calling.redux.Store
-import com.azure.android.communication.ui.calling.redux.action.LocalParticipantAction
-import com.azure.android.communication.ui.calling.redux.state.ReduxState
 import android.os.Bundle
-import com.azure.android.communication.ui.calling.redux.middleware.bluetooth.BluetoothDetector
-import com.azure.android.communication.ui.calling.redux.middleware.bluetooth.BluetoothDetectorImpl
-import com.azure.android.communication.ui.calling.redux.middleware.bluetooth.HeadsetDetector
+import com.azure.android.communication.ui.calling.utilities.audio.BluetoothDetector
+import com.azure.android.communication.ui.calling.utilities.audio.HeadsetDetector
 
 // Glues the Detector and Redux Together
 // Listens for Bluetooth devices and dispatches info
